requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cssselect==1.1.0
cycler==0.10.0
dateparser==1.0.0
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
GoogleNews==1.5.9
idna==3.2
importlib-metadata==4.6.4
importlib-resources==5.2.2
jieba3k==0.35.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.3.4
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
requests-file==1.5.1
scikit-learn==0.24.2
scipy==1.5.4
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
threadpoolctl==2.2.0
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.62.0
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
wordcloud==1.8.1
xlrd==2.0.1
zipp==3.5.0
